refactor(inline-edit): inline core into rhf/tsf wrappers, drop standalone item
- delete components/inline-edit; each wrapper is now self-contained (full duplication)
- add vertical/horizontal/no-label examples per wrapper
- remove the redundant submit button from examples (save-now persists per field)
