+++ apps/docs/content/docs/components/.gitignore
@@ -2,7 +2,6 @@
 confirmation-dialog.mdx
 copy-button.mdx
 hover-reveal.mdx
-inline-edit.mdx
 side-dialog.mdx
 submit-button.mdx
 theme-button.mdx
