--- packages/registry/items/components/inline-edit/index.mdx
@@ -1,72 +1,0 @@
----
-title: Inline Edit
-description: A click-to-edit value that saves on commit.
-registryName: inline-edit
----
-
-import { TypeTable } from 'fumadocs-ui/components/type-table';
-
-A value cell that displays text and turns into an editable input on click, saving on
-commit (Enter, or blur) through an async `onSave`. Read and edit states share one
-typography scale, so the value never jumps size — editing is seamless by default
-(`variant='ghost'`). Use `variant='boxed'` for a thin bordered input and `size='title'`
-for headings. For form integration with schema validation, use
-[`rhf-inline-edit`](/docs/react-hook-form/inline-edit) or
-[`tsf-inline-edit`](/docs/tanstack-form/inline-edit), which bind this cell to a form.
-
-## Examples
-
-<ItemExamples registryName={'inline-edit'} />
-
-## Props
-
-<TypeTable
-  type={{
-    value: {
-      description: 'The current value to display and edit.',
-      type: 'string',
-    },
-    onSave: {
-      description:
-        'Called with the new value on commit. May return a promise; the cell shows a saving state while it is pending and an inline error if it rejects.',
-      type: '(next: string) => Promise<void> | void',
-    },
-    input: {
-      description: 'Which built-in editor to render in edit mode.',
-      type: "'text' | 'textarea'",
-      default: "'text'",
-    },
-    variant: {
-      description: 'Edit-mode chrome. ghost is seamless (no border); boxed shows a thin bordered input.',
-      type: "'ghost' | 'boxed'",
-      default: "'ghost'",
-    },
-    size: {
-      description: 'Typography and spacing scale, shared by the read and edit states.',
-      type: "'sm' | 'default' | 'title'",
-      default: "'default'",
-    },
-    error: {
-      description: 'External error message to display (e.g. from a form validator). Shown in both read and edit modes.',
-      type: 'string',
-    },
-    placeholder: {
-      description: 'Shown when the value is empty, in read mode.',
-      type: 'string',
-    },
-    validate: {
-      description: 'Synchronous validation run before saving. Return a message to block the commit.',
-      type: '(next: string) => string | undefined',
-    },
-    canEdit: {
-      description: 'Whether clicking enters edit mode.',
-      type: 'boolean',
-      default: 'true',
-    },
-    children: {
-      description:
-        'Render-prop escape hatch for a custom editor. Receives value, setValue, commit, cancel, and the resolved className.',
-      type: 'function',
-    },
-  }}
-/>
