+++ packages/registry/items/react-hook-form/inline-edit/index.mdx
@@ -6,10 +6,11 @@ registryName: rhf-inline-edit
 
 import { TypeTable } from 'fumadocs-ui/components/type-table';
 
-Binds the [`inline-edit`](/docs/components/inline-edit) cell to React Hook Form. The value
-lives in the form (so it is covered by the form schema and submit), and on each commit the
-field is validated; when valid, an optional `onSave` persists it immediately — schema
-validation and save-now together. The validation error is surfaced inside the cell.
+A click-to-edit field for React Hook Form. The value displays as text and turns into an
+editable input on click; read and edit share one typography scale, so editing is seamless.
+The value lives in the form (covered by the form schema), and on each commit the field is
+validated; when valid, an optional `onSave` persists it immediately — schema validation and
+save-now together. The validation error is surfaced inside the cell.
 
 ## Examples
